cpufreq: impulse: reduce timer with screen off
Original patch for interactive by franciscofranco.

Standard Qualcomm configuration on userspace's PowerHAL is to increase the timer_rate value to reduce power consumption. That doesn't seem to happen with Angler (or any other Nexus for that matter) so this patch implements it in the governor driver instead

Tested with Google Play Music streaming in its higher bitrate, and there's no audio issues

Signed-off-by: franciscofranco <franciscofranco.1990@gmail.com>